cloudfunction/ms_auth_token/source/requirements.txt
google-auth==2.3.3
requests==2.27.1